made a new component for flat pages and updated the about page to pull from the cms
